backend/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.4
django-cors-headers==3.1.0
django-filter==2.2.0
django-rest-auth==0.9.5
django-simpleui==2.6
django-uuslug==1.1.8
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
idna==2.8
PyJWT==1.7.1
python-slugify==3.0.3
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
urllib3==1.25.3
